requirements.txt
flask
numpy
tensorflow==2.8.3
Pillow